read_from_gcs_filtering_push_to_mysql/requirements.txt
selenium
webdriver-manager
pandas
mysql-connector
gcloud
gcsfs